rest: split the internal router into one module per surface
The single internal router had grown to cover auth, ingest, usage and detectors
in one file. Each surface gets its own module behind the same package name, so a
change to one no longer means editing a file that already did too much. The
detector endpoint fixture follows the package the router now lives in.

(cherry picked from commit e02a735b5e4fbc07e02c8b7c8527e735d3d1a99d)
